SLVS-2104 Split AnalysisSettings to Global and Solution AnalysisSettings
Maintain the UserDefinedFileExclusions on the AnalysisSettings for the json and the FileExclusionsViewModel
