+++ src/Core/UserRuleSettings/AnalysisSettings.cs
@@ -20,80 +20,45 @@
 
 using System.Collections.Immutable;
 using System.IO;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using SonarLint.VisualStudio.Core.Helpers;
 
 namespace SonarLint.VisualStudio.Core.UserRuleSettings;
 
-/**
- * Example config file - same format as the VS Code settings.json file, with the addition of "parameters"
- * and "severity", both of which are optional.
- {
-     ...
-     "sonarlint.rules": {
-         "typescript:S2685": {
-             "level": "on"
-         },
-         "javascript:EqEqEq": {
-             "level": "on"
-         },
-         "cpp:S967": {
-             "level": "off"
-         },
-         "c:CommentedCode": {
-             "level": "on",
-             "Parameters": {
-               "key1": "value1",
-               "key2": "value2"
-             },
-             "severity": "Critical"
-         },
-     },
-     "sonarlint.analysisExcludesStandalone": "file1.cs,file2.cs"
-     ...
- }
- */
 public class AnalysisSettings
 {
     private const string AnyDirectoryWildcard = "**";
     private static readonly string AnyRootPrefix = AnyDirectoryWildcard + Path.AltDirectorySeparatorChar;
     private readonly ImmutableArray<string> userDefinedFileExclusions = ImmutableArray<string>.Empty;
 
-    [JsonProperty("sonarlint.rules")]
-    [JsonConverter(typeof(ImmutableDictionaryIgnoreCaseConverter<string, RuleConfig>))]
-    public ImmutableDictionary<string, RuleConfig> Rules { get; init; }
+    public ImmutableDictionary<string, RuleConfig> Rules { get; }
 
-    [JsonIgnore] // todo https://sonarsource.atlassian.net/browse/SLVS-2059
-    public ImmutableDictionary<string, string> AnalysisProperties { get; init; } = ImmutableDictionary<string, string>.Empty;
+    public ImmutableDictionary<string, string> AnalysisProperties { get; }
 
-    [JsonProperty("sonarlint.analysisExcludesStandalone")]
-    [JsonConverter(typeof(CommaSeparatedStringArrayConverter))]
     public ImmutableArray<string> UserDefinedFileExclusions
     {
         get => userDefinedFileExclusions;
-        init
+        private init
         {
-            userDefinedFileExclusions = value;
-            NormalizedFileExclusions = value.Select(NormalizePath).ToArray();
+            userDefinedFileExclusions = value
+                .Where(path => !string.IsNullOrWhiteSpace(path))
+                .ToImmutableArray();
+            NormalizedFileExclusions = userDefinedFileExclusions
+                .Select(NormalizePath)
+                .ToImmutableArray();
         }
     }
 
-    [JsonIgnore]
-    public string[] NormalizedFileExclusions { get; private init; } = [];
+    public ImmutableArray<string> NormalizedFileExclusions { get; private init; }
 
-    public AnalysisSettings(Dictionary<string, RuleConfig> rules, IEnumerable<string> fileExclusions)
+    public AnalysisSettings(ImmutableDictionary<string, RuleConfig> rules = null, IEnumerable<string> fileExclusions = null, ImmutableDictionary<string, string> analysisProperties = null)
     {
-        Rules = rules.ToImmutableDictionary(StringComparer.OrdinalIgnoreCase);
-        UserDefinedFileExclusions = fileExclusions.ToImmutableArray();
+        Rules = rules ?? ImmutableDictionary<string, RuleConfig>.Empty;
+        UserDefinedFileExclusions = fileExclusions?.ToImmutableArray() ?? ImmutableArray<string>.Empty;
+        AnalysisProperties = analysisProperties ?? ImmutableDictionary<string, string>.Empty;
     }
 
-    public AnalysisSettings(ImmutableDictionary<string, RuleConfig> rules, IEnumerable<string> fileExclusions) : this(rules.ToDictionary(kvp => kvp.Key, kvp => kvp.Value), fileExclusions) { }
-
-    public AnalysisSettings()
+    public AnalysisSettings(Dictionary<string, RuleConfig> rules, IEnumerable<string> fileExclusions, Dictionary<string, string> analysisProperties = null) : this(
+        rules.ToImmutableDictionary(StringComparer.OrdinalIgnoreCase), fileExclusions, analysisProperties?.ToImmutableDictionary(StringComparer.OrdinalIgnoreCase))
     {
-        Rules = ImmutableDictionary.Create<string, RuleConfig>(StringComparer.OrdinalIgnoreCase);
-        UserDefinedFileExclusions = ImmutableArray<string>.Empty;
     }
 
     private static string NormalizePath(string path)
@@ -120,29 +85,4 @@ public class AnalysisSettings
         var invalidChars = Path.GetInvalidPathChars();
         return path.IndexOfAny(invalidChars) >= 0;
     }
-}
-
-[method: JsonConstructor]
-public class RuleConfig(RuleLevel level, Dictionary<string, string> parameters)
-{
-    public RuleConfig(RuleLevel level) : this(level, null) { }
-
-    [JsonProperty("level")]
-    [JsonConverter(typeof(StringEnumConverter))]
-    public RuleLevel Level { get; init; } = level;
-
-    // Note: property will be null if "parameters" is missing from the file.
-    // This is what we want: most rules won't have parameters and we want to avoid
-    // creating hundreds of unnecessary empty dictionaries.
-    // The only downside is that the dictionary that is created will use the default
-    // comparer, which is case-sensitive.
-    [JsonProperty("parameters", NullValueHandling = NullValueHandling.Ignore)]
-    [JsonConverter(typeof(ImmutableDictionaryIgnoreCaseConverter<string, string>))]
-    public ImmutableDictionary<string, string> Parameters { get; init; } = parameters?.ToImmutableDictionary(StringComparer.OrdinalIgnoreCase);
-}
-
-public enum RuleLevel
-{
-    On,
-    Off
 }
