+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithStrongNames.txt
@@ -1,7 +1,7 @@
 ﻿---
 ################################
 # Assembly references report
-# Report date/time: 2025-04-28T15:02:40.7323621Z
+# Report date/time: 2025-05-01T12:30:43.4182123Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.11.0.223
@@ -331,9 +331,10 @@ Referenced assemblies:
 - 'SonarLint.VisualStudio.IssueVisualization.Security, Version=8.18.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'SonarLint.VisualStudio.SLCore, Version=8.18.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
 - 'System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
+- 'System.Collections.Immutable, Version=5.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'
 - 'System.ComponentModel.Composition, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
 - 'System.Core, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089'
-# Number of references: 9
+# Number of references: 10
 
 ---
 Assembly: 'SonarQube.Client, Version=8.18.0.0, Culture=neutral, PublicKeyToken=c5b62af9de6d7244'
